+++ test/webapi/test_minheap.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import numpy as np
 from numpy.testing import assert_almost_equal
 
-from cate.util.minheap import MinHeap
+from cate.webapi.minheap import MinHeap
 
 
 class MinHeapTest(TestCase):
